+++ src/components/ids-data-grid/demos/infinite-scroll.ts
@@ -4,7 +4,7 @@ import type { IdsDataGridColumn } from '../ids-data-grid-column';
 import productsJSON from '../../../assets/data/products.json';
 
 // Example for populating the DataGrid
-const dataGrid = document.querySelector<IdsDataGrid>('#data-grid-virtual-scroll')!;
+const dataGrid = document.querySelector<IdsDataGrid>('#data-grid-infinite-scroll')!;
 
 // Do an ajax request
 const url: any = productsJSON;
@@ -80,22 +80,29 @@ columns.push({
 
 dataGrid.columns = columns;
 
-const setData = async () => {
+const fetchData = async (startIndex = 0) => {
   const res = await fetch(url);
   const data = await res.json();
-  dataGrid.data = data;
+  return data.splice(startIndex, 33);
+};
+
+const setData = async () => {
+  dataGrid.data = await fetchData();
 };
 
 setData();
 
-dataGrid.addEventListener('selectionchanged', (e: Event) => {
-  console.info(`Selection Changed`, (<CustomEvent>e).detail);
-});
-
 dataGrid.addEventListener('scrollstart', async (e: Event) => {
-  console.info(`Virtual Scroll reached start`, (<CustomEvent>e).detail);
+  console.info(`Scroll reached start`, (<CustomEvent>e).detail);
 });
 
 dataGrid.addEventListener('scrollend', async (e: Event) => {
-  console.info(`Virtual Scroll reached end`, (<CustomEvent>e).detail);
+  console.info(`Scroll reached end`, (<CustomEvent>e).detail);
+  const endIndex = (<CustomEvent>e).detail?.value || 0;
+  const MAX_ROW_INDEX = 299;
+
+  if (endIndex < MAX_ROW_INDEX) {
+    const moreData = await fetchData(endIndex + 1);
+    dataGrid.appendData(moreData);
+  }
 });