+++ src/components/ids-data-grid/ids-data-grid.ts
@@ -9,7 +9,7 @@ import IdsDataSource from '../../core/ids-data-source';
 import IdsDataGridFormatters from './ids-data-grid-formatters';
 import { editors } from './ids-data-grid-editors';
 import IdsDataGridFilters, { IdsDataGridFilterConditions } from './ids-data-grid-filters';
-import { IdsDataGridContextmenuArgs, setContextmenu } from './ids-data-grid-contextmenu';
+import { IdsDataGridContextmenuArgs, setContextmenu, getContextmenuElem } from './ids-data-grid-contextmenu';
 import { IdsDataGridColumn, IdsDataGridColumnGroup } from './ids-data-grid-column';
 
 import IdsPopupMenu from '../ids-popup-menu/ids-popup-menu';
@@ -19,7 +19,7 @@ import {
   resetEmptyMessageElements,
   hideEmptyMessage,
   IdsDataGridToggleEmptyMessage,
-  emptyMessageTemplate,
+  emptyMessageTemplate
 } from './ids-data-grid-empty-message';
 
 // Styles
@@ -93,6 +93,19 @@ export default class IdsDataGrid extends Base {
 
   cacheHash = Math.random().toString(32).substring(2, 10);
 
+  /**
+   * Types for contextmenu.
+   * @private
+   */
+  contextmenuTypes = {
+    BODY_CELL: 'body-cell',
+    BODY_CELL_EDITOR: 'body-cell-editor',
+    HEADER_TITLE: 'header-title',
+    HEADER_ICON: 'header-icon',
+    HEADER_FILTER: 'header-filter',
+    HEADER_FILTER_BUTTON: 'header-filter-button',
+  };
+
   constructor() {
     super();
 
@@ -129,7 +142,7 @@ export default class IdsDataGrid extends Base {
 
     super.connectedCallback();
     this.redrawBody();
-    this.#attachVirtualScrollEvent();
+    this.#attachScrollEvents();
   }
 
   /** Reference to datasource API */
@@ -206,6 +219,7 @@ export default class IdsDataGrid extends Base {
 
     let cssClasses = `${this.alternateRowShading ? ' alt-row-shading' : ''}`;
     cssClasses += `${this.listStyle ? ' is-list-style' : ''}`;
+    const emptyMesageTemplate = emptyMessageTemplate.apply(this);
 
     const html = `<div class="ids-data-grid-wrapper">
         <span class="ids-data-grid-sort-arrows"></span>
@@ -213,6 +227,7 @@ export default class IdsDataGrid extends Base {
           ${IdsDataGridHeader.template(this)}
           ${this.bodyTemplate()}
         </div>
+        ${emptyMesageTemplate}
         <slot name="menu-container"></slot>
         <slot name="contextmenu"></slot>
         <slot name="header-contextmenu"></slot>
@@ -303,7 +318,7 @@ export default class IdsDataGrid extends Base {
     this.header.setHeaderCheckbox();
     this.#attachEventHandlers();
     this.#attachKeyboardListeners();
-    this.#attachVirtualScrollEvent();
+    this.#attachScrollEvents();
     this.setupTooltip();
 
     // Attach post filters setting
@@ -336,11 +351,17 @@ export default class IdsDataGrid extends Base {
    * Contextmenu stuff use for info and events
    * @private
    */
-  contextMenuInfo: {
+  contextmenuInfo: {
     menu?: IdsPopupMenu,
     target?: HTMLElement,
     callbackArgs?: IdsDataGridContextmenuArgs
   } = {};
+
+  /**
+   * Track contextmenu data dynamicly changed by the user.
+   * @private
+   */
+  isDynamicContextmenu = false;
 
   /**
    * Body template markup
@@ -348,9 +369,7 @@ export default class IdsDataGrid extends Base {
    * @returns {string} The template
    */
   bodyTemplate() {
-    const emptyMesageTemplate = emptyMessageTemplate.apply(this);
-
-    return `${emptyMesageTemplate}<div class="ids-data-grid-body" part="contents" role="rowgroup">${this.bodyInnerTemplate()}</div>`;
+    return `<div class="ids-data-grid-body" part="contents" role="rowgroup">${this.bodyInnerTemplate()}</div>`;
   }
 
   /**
@@ -1039,6 +1058,20 @@ export default class IdsDataGrid extends Base {
   get columnGroups() { return this.state?.columnsGroups || null; }
 
   /**
+   * Use this to add more data to the datagrid's existing dataset.
+   * This will automatically render additional rows in the datagrid.
+   * @param {Array} value The array to use
+   */
+  appendData(value: Array<Record<string, any>>) {
+    if (this.virtualScroll) {
+      this.datasource.data = this.data.concat(value);
+    } else {
+      this.data = this.data.concat(value);
+    }
+    this.#attachScrollEvents();
+  }
+
+  /**
    * Set the data of the data grid
    * @param {Array} value The array to use
    */
@@ -1128,7 +1161,12 @@ export default class IdsDataGrid extends Base {
    */
   set headerMenuData(value) {
     this.header.state.headerMenuData = value;
-    setContextmenu.apply(this);
+    if (!this.isDynamicContextmenu) {
+      const headerMenu: any = getContextmenuElem.apply(this, [true]);
+      if (headerMenu) headerMenu.data = value;
+      setContextmenu.apply(this);
+    }
+    this.isDynamicContextmenu = false;
   }
 
   get headerMenuData() { return this.header.state.headerMenuData; }
@@ -1153,7 +1191,12 @@ export default class IdsDataGrid extends Base {
    */
   set menuData(value) {
     this.state.menuData = value;
-    setContextmenu.apply(this);
+    if (!this.isDynamicContextmenu) {
+      const menu: any = getContextmenuElem.apply(this);
+      if (menu) menu.data = value;
+      setContextmenu.apply(this);
+    }
+    this.isDynamicContextmenu = false;
   }
 
   get menuData() { return this?.state?.menuData; }
@@ -1199,9 +1242,46 @@ export default class IdsDataGrid extends Base {
     };
   }
 
-  /* Attach Events for virtual scrolling */
+  /* Attach Events for global scrolling */
+  #attachScrollEvents() {
+    let debounceRowIndex = 0;
+    this.offEvent('scroll.data-grid', this.container);
+    this.onEvent('scroll.data-grid', this.container, () => {
+      const virtualScrollSettings = this.virtualScrollSettings;
+      const scrollTop = this.container!.scrollTop;
+      const clientHeight = this.container!.clientHeight;
+
+      const rowIndex = Math.floor(scrollTop / virtualScrollSettings.ROW_HEIGHT);
+
+      if (rowIndex === debounceRowIndex) return;
+      debounceRowIndex = rowIndex;
+
+      const data = this.data;
+      const rows = this.rows;
+      const maxHeight = virtualScrollSettings.ROW_HEIGHT * data.length;
+
+      const reachedTheTop = rowIndex <= 0;
+      const reachedTheBottom = (scrollTop + clientHeight) >= maxHeight;
+
+      if (reachedTheTop) {
+        const firstRow: any = rows[0];
+        this.#triggerCustomScrollEvent(firstRow.rowIndex, 'start');
+      }
+      if (reachedTheBottom) {
+        const lastRow: any = rows[rows.length - 1];
+        this.#triggerCustomScrollEvent(lastRow.rowIndex, 'end');
+      }
+      if (!reachedTheTop && !reachedTheBottom) {
+        this.#triggerCustomScrollEvent(0);
+      }
+    }, { capture: true, passive: true }); // @see https://javascript.info/bubbling-and-capturing#capturing
+
+    this.#attachVirtualScrollEvent();
+  }
+
   #attachVirtualScrollEvent() {
     if (!this.virtualScroll) return;
+
     const virtualScrollSettings = this.virtualScrollSettings;
     const data = this.data;
 
@@ -1224,6 +1304,22 @@ export default class IdsDataGrid extends Base {
 
       this.scrollRowIntoView(rowIndex, false);
     }, { capture: true, passive: true }); // @see https://javascript.info/bubbling-and-capturing#capturing
+  }
+
+  #customScrollEventCache: { [key: string]: number } = {};
+
+  #triggerCustomScrollEvent(rowIndex: number, eventType?: 'start' | 'end') {
+    if (!eventType) {
+      this.#customScrollEventCache = {}; // reset event-cache
+    } else if (rowIndex !== this.#customScrollEventCache[eventType]) {
+      this.#customScrollEventCache[eventType] = rowIndex;
+
+      this.triggerEvent(`scroll${eventType}`, this, {
+        bubbles: true,
+        composed: true,
+        detail: { elem: this, value: rowIndex }
+      });
+    }
   }
 
   /**
@@ -1301,6 +1397,8 @@ export default class IdsDataGrid extends Base {
     bufferRowIndex = Math.min(bufferRowIndex, maxRowIndex);
 
     if (isInRange) {
+      // if rowIndex is in range of the currently visible rows:
+      // then we should only move rows up or down according to how big the buffer should be.
       const moveRowsDown = bufferRowIndex - firstRowIndex;
       const moveRowsUp = Math.abs(moveRowsDown);
 
@@ -1314,6 +1412,8 @@ export default class IdsDataGrid extends Base {
         return; // exit early because nothing to do.
       }
     } else if (isAboveFirstRow) {
+      // if rowIndex should appear above the currently visible rows,
+      // then we must figure out how many rows we must move up from the bottom to render the rowIndex row
       const moveRowsUp = Math.abs(bufferRowIndex - firstRowIndex);
 
       if (moveRowsUp < virtualScrollSettings.NUM_ROWS) {
@@ -1322,6 +1422,8 @@ export default class IdsDataGrid extends Base {
         this.#recycleAllRows(bufferRowIndex);
       }
     } else if (isBelowLastRow) {
+      // if rowIndex should appear below the currently visible rows,
+      // then we recycle all rows, since none of the visible rows are needed
       this.#recycleAllRows(bufferRowIndex);
     }
 
@@ -1344,7 +1446,7 @@ export default class IdsDataGrid extends Base {
     });
   }
 
-  /* Recycle the rows duing scrolling */
+  /* Recycle the rows during scrolling */
   #recycleAllRows(topRowIndex: number) {
     const rows = this.rows;
     if (!rows.length) return;
@@ -1366,7 +1468,7 @@ export default class IdsDataGrid extends Base {
     });
   }
 
-  /* Recycle the rows duing scrolling from the top */
+  /* Recycle the rows during scrolling from the top */
   #recycleTopRowsDown(rowCount: number) {
     const rows = this.rows;
     if (!rowCount || !rows.length) return;
@@ -1397,7 +1499,7 @@ export default class IdsDataGrid extends Base {
     });
   }
 
-  /* Recycle the rows duing scrolling from the bottom */
+  /* Recycle the rows during scrolling from the bottom */
   #recycleBottomRowsUp(rowCount: number) {
     const rows = this.rows;
     if (!rowCount || !rows.length) return;