+++ src/core/ids-attributes.ts
@@ -227,6 +227,7 @@ export const attributes = {
   MAX_FILES: 'max-files',
   MAX_FILES_IN_PROCESS: 'max-files-in-process',
   MAX_FILE_SIZE: 'max-file-size',
+  MAX_HEIGHT: 'max-height',
   MAX_TRANSFORM_X: 'max-transform-x',
   MAX_TRANSFORM_Y: 'max-transform-y',
   MENU: 'menu',
@@ -466,7 +467,7 @@ export const prefix = {
   PREFIX: 'ids'
 };
 
-export const version = '1.0.0-beta.6';
+export const version = '1.0.0-beta.7';
 
 export enum IdsDirection {
   Up = 'up',